# docs/openapi.yaml
openapi: 3.0.3
info:
  title: BridgeCore Platform API
  description: >
    BridgeCore is a reusable multi-tenant backend platform providing
    authentication, authorization, feature entitlements, API keys, usage
    metering, and audit logging as REST APIs for downstream products.
  version: "0.1.0"
  contact:
    name: BridgeCore Platform Team
servers:
  - url: http://localhost:8080
    description: Local development
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
...
paths:
...
  /api/v1/tenants:
    post:
      tags: [tenants]
      summary: Create a tenant (admin only)
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, slug]
              properties:
                name: { type: string }
                slug: { type: string }
                plan: { type: string, enum: [free, pro, enterprise] }
      responses:
        "201": { description: Created }
    get:
      tags: [tenants]
      summary: List tenants (admin only)
      security: [{ BearerAuth: [] }]
      parameters:
        - in: query
          name: search
          schema: { type: string }
        - in: query
          name: page
          schema: { type: integer }
        - in: query
          name: page_size
          schema: { type: integer }
      responses:
        "200": { description: OK }
  /api/v1/tenants/{id}:
    get:
      tags: [tenants]
      summary: Get a tenant by ID
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200": { description: OK }
        "404": { description: Not found }
    put:
      tags: [tenants]
      summary: Update a tenant (admin only)
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200": { description: OK }
    delete:
      tags: [tenants]
      summary: Soft-delete a tenant (admin only)
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200": { description: OK }
  /api/v1/users:
    get:
      tags: [users]
      summary: List users in the current tenant
      security: [{ BearerAuth: [] }]
      responses:
        "200": { description: OK }
  /api/v1/users/{id}/role:
    patch:
      tags: [users]
      summary: Change a user's RBAC role (admin only)
      security: [{ BearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role: { type: string, enum: [admin, developer, viewer] }
      responses:
        "200": { description: OK }
...
  /api/v1/usage:
    get:
      tags: [usage]
      summary: List raw usage records for the current tenant
      security: [{ BearerAuth: [], ApiKeyAuth: [] }]
      parameters:
        - in: query
          name: endpoint
          schema: { type: string }
        - in: query
          name: method
          schema: { type: string }
        - in: query
          name: from
          schema: { type: string, format: date-time }
        - in: query
          name: to
          schema: { type: string, format: date-time }
        - in: query
          name: page
          schema: { type: integer }
        - in: query
          name: page_size
          schema: { type: integer }
      responses:
        "200": { description: OK }
  /api/v1/usage/summary:
    get:
      tags: [usage]
      summary: Get aggregated usage summary per endpoint
      security: [{ BearerAuth: [], ApiKeyAuth: [] }]
      responses:
        "200": { description: OK }
  /api/v1/usage/export:
    get:
      tags: [usage]
      summary: >
        Export usage records as CSV. Requires the "usage.export" feature
        entitlement (Pro/Enterprise plans) — gated by middleware before
        the handler runs.
      security: [{ BearerAuth: [], ApiKeyAuth: [] }]
      responses:
        "200":
          description: CSV file
          content:
            text/csv:
              schema: { type: string }
        "403": { description: Plan does not include this feature }
...